Remove the hspace and vspace mxins and add a general margin/padding mixin
Decided to remove the hspace() and vspace() mixin. The problem they
created was that you need both to set all or three margins. Another
problem was that it creates unnecessary errors with declaration order.
So I created a simple single mixin to handle adding padding or margin
based on a single value, list or even a map.