Change rem-font-size() mixin to a general font-size() mixin
The rem-font-size mixin only returns a rem font-size with a px fallback
value. Decided to opt for a setup where we specify the unit in a config
along with whether or not we use px fallbacks. This way the mixin can
always be used for setting the font-size regardless if you use rem
units or not.