Move default animation flags to own file and add keyframe animations to mixin
Initial setup with both keyframes and mixin wrapped in a flag conditional
would not work it seemed. Because of the execution order the default value
would always take precedence. Moving the keyframe to the mixin fixes that.
And using the @at-root control allows for correct placement of the keyframe
afterwards when compiling.